moving sphere
A sphere that is travelling linearly between the position (x1, y1, z1)
and (x2, y2, z2) while the scene is being rendered. Because of this
motion, the sphere should be motion blurred. Each ray that is tested
against this sphere should be assigned a random time in the interval
[0,1], and the sphere's position should be determined by this random
value.