FIXED ACNE
Fixed Acne by moving the shadow rays up a little. This is happening
because for a given shadow ray, it calls intersection on the ShapeList
which then returns an IntersectionObject. The pointer the the actual
object this intersects with however is the same so then we disregard it
and then we do not query for another one. This makes the surface acne
"backwards" so it is lighter where it should be dark because we are
disregarding this intersection altogether and do not look for another
one.